pyproject.toml
[tool.poetry]
name = "alignment-ecommerce"
version = "0.1.0"
description = ""
authors = ["Mark Bruderer <markbruderer123@gmail.com>"]
readme = "README.md"
packages = [{ include = "alignment_ecommerce" }]

[tool.poetry.scripts]
dev = "streamlit run app.py -- debug"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.3.9"
humanloop = "^0.5.18"
elasticsearch = "^8.11.1"
streamlit = "^1.29.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
